refactor(experiments): move script tests adjacent to sources, run with vitest
