+++ claude-plugins/experiments/scripts/lib/cache.test.mjs
@@ -2,7 +2,7 @@ import assert from "node:assert/strict";
 import { mkdtempSync, utimesSync, writeFileSync } from "node:fs";
 import { tmpdir } from "node:os";
 import { join } from "node:path";
-import { test } from "node:test";
+import { test } from "vitest";
 import {
     ensurePackageMeta,
     listCachedVersions,
@@ -19,7 +19,7 @@ import {
     writeFailedVersion,
     writeRawSource,
     writeVerifiedVersion,
-} from "../lib/cache.mjs";
+} from "./cache.mjs";
 
 function tempRoot() {
     return mkdtempSync(join(tmpdir(), "chlog-cache-"));
