+++ claude-plugins/experiments/vitest.config.mjs
@@ -1,0 +1,8 @@
+import { configDefaults, defineConfig } from "vitest/config";
+
+export default defineConfig({
+    test: {
+        include: ["scripts/**/*.test.mjs"],
+        exclude: [...configDefaults.exclude, "**/.stryker-tmp/**"],
+    },
+});
